Separate "pure" expressions from user expressions
We now support lambda expressions without bodies, which are useful as an
intermediate step while constructing an expression. They render with a red box
where the body should go. This required changing the model code so that we have
two types of AST, one with pure lambda calculus expressions and one with a
user-facing expression that may not be pure. In addition to a missing lambda
body, it might also have shorthands (e.g. Y instead of the full Y combinator)
and maybe custom colors or things like that.