FRPize more drag events
This introduces horrible bugs where objects become undraggable, but I'll fix
that later. We now go through multiple transformations to get an observable that
produces independent drag operations, which themselves are observables of
points. That means that the subscriber no longer needs to think about which
pointer is the active one, and can just ignore pointers that come in while one
is active.