Refactor things to use views
Instead of dealing with native views, we now have wrappers for the different
types of native views, which should make it easier to do operations like "listen
to drag operations on the parameter view in the lambda".

Plenty of law of demeter violations currently, but the idea is that views should
control all Android-specific stuff, and manage the visual changes during drags
and things like that.