Avoid accumulating more and more DragSources over time
Each drag source now is given its rx subscription when the event starts, and it
can use that to indicate that the drag source should go away. For example, we do
this when dragging a lambda body out of a lambda (since it gets replaced with a
top-level drag source), but not when creating a variable from a lambda
parameter.