Fix bug where lambdas would drop into themselves
If we moved a lambda expression containing containing a lambda body drop target,
it would match itself, and doing a drop operation would make it its own parent,
so it would just disappear. Fixed by changing the hit test logic to specifically
ignore the expression being dragged.

Ideally, we would do all of this independent of Android logic, but it's pretty
easy to do a parent test on the native views, so I just did that.