Only start drag operations on down events
Like a fool, I ignored the part of the Android docs where it mentions that the
down, move, and up motion events aren't guaranteed to be consistent. It turns
out that they weren't kidding, I was starting drag events way before they
actually occurred, which cased the offset to be incorrect. Fixed by only
starting a drag event when the action type is DOWN.