Don't use weak references in drag tracking
I'm not sure if this was actually causing problems, but I saw problems
correlated with GC, so I'm just being safe for now. There are enough other
memory leaks that I doubt this makes things any worse.

Also make the screen empty on startup to make things a little easier to test.