Fix bug where func views would steal click events
For example, a function x(x) would have a view for the left x and a view for the
right x, and the left view would cover the function view itself. Ideally, we
would be smart enough to forward the events in that situation, but the hack I'm
using for now is to use Android's ability to disable views, which causes all
mouse events to the view to be ignored.