Allow dropping variables back on parameters
It's fairly common to accidentlly tap the parameter of a lambda and create a new
variable, and before this commit the only way to actually get rid of it was to
drag it to the right side of the screen. Now, you can just "put it back" and it
will go away.