Fix visual bug when expressions are halfway off the bottom or right
The problem here was that the RelativeLayout parent seems to have a rule that
children that are partway off of the layout are given an AT_MOST measure spec
(but only on the bottom or right side, not the top or left side). This caused
expressions to look weird when they were partway off the edge of the screen. My
solution, which is a bit of a hack, but works, is to define a custom layout that
ignores the incoming measure spec and just uses UNSPECIFIED. This is effectively
a way to say that expressions should ignore any input from their enclosing
container when deciding their size.

Also do some other minor cleanups.