Rework palette styles
Instead of putting it in a hard-coded position designed for landscape Nexus 9,
we now align it to the right of the screen and put it in a ScrollView. I also
had to do some tricks to get the ScrollView to behave properly: ignoring
anything from onInterceptTouchEvent to basically dictate that scrolling should
only happen directly on the palette.