Change the palette to be a right-side drawer
I also refactored a lot of code to rely more on XML instead of constructing
everything programmatically.

This required some hacks/tweaks where we ignore a bunch of touch events so that
the user can interact with things while the palette is open. Maybe in the future
I'll be able to make it smarter and make it possible to dismiss or scroll the
palette in a more intuitive way by recognizing the gesture.