Don't allow expression deletion when the palette is closed
Previously it would delete expressions that were dragged to the right side of
the screen, but this was annoying. Now it's much more comfortable to work with
expressions because you can just close the palette and not worry about
accidentally deleting things, and panning lets you work with larger expressions.