Add brackets and parens to make the grouping more clear
This takes away some of the magic because you're no longer just rearranging
cards, but it looks like it makes the grouping way more clear and makes it so
you don't need as much precision to place a lambda in the right place. It looks
a lot nicer when the parens and brackets stretch to fit their height, but that's
hard to do with regular text, it seems, so I used ImageViews instead, since
those can be stretched programmatically.

Since I'm using images, it's easiest to just make them black, so for now all
text is black. But in the future it may be nice to switch back to the dark gray
color that it was before (or whatever is in the theme).