Change styleLayout to be an internal function in the renderer
This causes a regression, which will be fixed in the next commit: when creating
a new view like a function call, we end up re-styling the top-level view from
being in a realtive layout to being in a linear layout, which is why we needed
to call this style function. I think a way to fix this is to make another
"container" view in the view hierarchy for top-level expressions. The container
view always lives in a relative layout, and normal expressions always will live
in a linear layout.