Fix display bug when creating a function call at the top level
When expressions switch between top-level expressions and non-top-level
expressions, their LayoutParams need to change because they're living in a
different layout. When a normal expression was added to the top level and then
moved back, it would lose its styles, which caused a display bug where
expressions would have border size 0. The old way to fix this was to have code
that would re-clobber the LayoutParams with the normal ones for expressions when
it changes to not be a top-level expression.

The new approach is more careful: we keep track of the original LayoutParams
when an expression switches to being top-level, then restore it when we detach
it. However, this required another change to make us more disciplined about how
we modify the view hierarchy. Previously, we would eagerly (and redundantly)
detach the left-side view when creating a function call, to avoid a view with
duplicate parents, but that messed this case up. Now, we only detach things that
are already attached, and we shift the order of function call creation so that
we first detach, then reparent, then attach.

It's more code everywhere, but should be easier to reason about and is a cleaner
solution since there isn't any code reaching into the view hierarchy to modify
it.