Set min API level to 15 and fix compat crashers
This mostly involved creating a new Compat class where we keep all of our
version checks. This ended up not being so bad. There are still some visual
issues on other versions/themes, so I'll need to figure that out in a later
commit, but the app certainly works on old versions with this commit.

I only tested on 16 (genymotion doesn't have images for API 15), but lint
doesn't complain, so I might as well throw it in.