Make new expressions appear above the palette for the initial drag
This required restructuring the view hierarchy into a normal "canvas root" and a
special "above palette root". Expressions from the palette start their life in
the above palette root, then move to the canvas root after being dropped for the
first time. This was done by passing around both RelativeLayouts and doing a
special detach operation on the first endDrag.