Set up ProGuard
I have it set to minify but not obfuscate in both debug and release. This should
avoid nasty broken-only-in-release bugs and avoid some of the debugging
headaches proguard can have. The APK size is now about 1.2MB, and it would be
950KB with full obfuscation and optimization, which doesn't seem worth it.

This also required upgrading RxJava. I also removed the RxAndroid dependency,
since it turned out I wasn't using it.