Fix execute button positioning on pre-Lollipop devices
The old FAB implementation expands the button size so that it can draw a shadow,
which was messing with my positioning code. Now, the code uses the center and dp
measurements rather than relying on the width, height, or top-left of the FAB.

A related problem was that the hit box was much bigger for pre-Lollipop FAB. To
fix this, I wrote a custom dispatchTouchEvent implementation that checks the
button radius and only deals with events that are within that radius from the
center of the button.