Add first pass at reference expressions
It's impossible to create them for now, but the implementation was pleasantly
simple. They behave just like normal variables, but when converting to full
lambda expressions they expand to more than just lambda variables. In the
future, there will be a system for keeping track of definitions and eventually
modifying them.