Refactor dragging to use a generic DragData class
Instead of dragging always having an expression as its data, there's now an
interface for it, which can later be used for definitions.

To make it easier to implement drop target, a drop target can have a type
parameter which is the subclass of DragData that it accepts. This will be useful
because definitions won't be able to be dropped in many places.