Fix some more evaluator problems
We now are more careful about what to evaluate and what not to evaluate. In a
function call, we only try evaluating the right side if the left side looks like
it isn't going to change anymore. That determination is done by keeping track of
which unbound variables are "fixed", i.e. they are from lambdas that will never
progress.

Also, there are cases where varMarkers might actually be reused; they are
generated at compile time and correspond to a variable in the original code, but
we might have nested functions that have the same source function, which can
break substitutions. To fix this for now, I don't traverse into a lambda
matching the current varMarker when doing substitution. But that may be just
papering over the problem and a better solution may be a good idea in the long
run.