Implement lambda creation
It works pretty much like definition creation. Maybe the code could be shared,
but it's probably fine to have this duplication for now.

I also changed the dagger mechanism to allow the fragment to inject things like
the point converter. In the future it may be good to push fragment logic into
injected classes, but this commit certainly works.