Overhaul app structure and add dummy definition palette
Now all ButterKnife bindings are done directly in the dagger module. This
might be a pain if we end up wanting multiple screens, but it's pretty nice for
now and avoids a lot of the boilerplate of passing views around. Rather than
making views injected, they're just fields that can be accessed, for now, which
works since we don't use @Inject anywhere and everything is in one big module.

I also created a simple definition palette using nested DrawerLayouts and moved
the DrawerLayout code into its own class. Soon, it will get more interesting,
since it'll need to manage two drawers and the corresponding transitions.