Polish drawer animation
Now there's an underlying state independent of the UI that says which drawer, if
any, should be open, and any user operation might change that state and we do
whatever is necessary to move the current UI to that state. For example, if the
user double-taps a drawer button when it's closed, it will switch from opening
to closing even if it never fully opened.

This also required some details where we add a listener that will open the other
drawer when the current drawer is halfway closed, and we need to remove that
listener when performing another user operation. This prevents a subtle type of
bug where a listener can live longer than expected.