Fix bug where sometimes a reference would stay invalid
There's an ordering issue when decomposing a function call. We propagate
expression changes before re-attaching the new top-level element, but that means
that that top-level element misses the invalidate call, which sometimes is
necessary. In particular starting with "FOO := x(FOO)" and pulling out the FOO
would make the FOO definition no longer circular, but the FOO reference would
stay red. This fixes the issue by just invalidating references whenever
something moves to the top level.