Remove old stepping evaluator, optimize canStep
We used to do the dumb thing for canStep: just actually step the expression and
see if we get something. Now we instead use a simple recursive walk to see if
there are any lambdas on the left side of a function call, which is much faster
in practice. This was the last use of the old evaulator, so we can now just get
rid of it.