Fix some display issues with elevation animation when dragging
We were animating the translation z to 15dp, then to 5dp, but really, we wanted
the effective elevation to switch between those two values. So the translation
value should go between 10dp and 0dp, since that's the offset applied to the
normal elevation. This avoids a situation where the execute button would be too
low or where expressions would end up above the delete bar.