Change background-setting to always use drawables
This fixes a bug on pre-lollipop devices where we would set the background color
to white in some code paths, but really we should set it to the expression
background, which on old devices has a border to account for the fact that there
are no shadows.

This should also make it a little easier to improve the styling for highlighting
at some point.