Implement basic serialization and deserialization
I also added a Redux middleware to serialize all objects and deserialize them
from the reducer. The basic idea behind deserialization is that every object
that's a child of a value type is either another value type or something that's
already JSON-serializable. This actually misses the case of maps, but I can
hopefully add those later.